Updated files for nutrition tab and fitness tab
Deleted unused files and folders, updated the routes accordingly.

Renamed Fitness.js to EquipmentFilter.js and Dummy.js to MuscleFilter.js, updated header to show correct thing.

Added FitnessGoals.js and NutritionTips.js for the nutrition tab and followed same style as  the fitness tab.

Changed the prop id to pass in the title of the pages(instead of "/title"), Each function now uses usePath and updates the URL accordingly

Removed the excerpt tag from the database files and edited the handle accordingly for url/usePath